requirements.txt
requests==2.31.0
pyyaml==6.0.2
toml==0.10.2
python-dotenv==1.1.0
colorama==0.4.6
prompt_toolkit==3.0.41 